Remove gcc warning about redefinition of putchar with different visibility
In NDK bionic's stdio.h, putchar appears as both a function prototype and a macro
(later in the same header).  Because it's defined as a macro, libc++ replaces
it with a function but with *different* visibility.  GCC 4.x complains [-Wattributes]

   include/stdio.h:236:5: warning: conflicts with previous declaration here [-Wattributes]
   int  putchar(int);

Undefine putchar to avoid redefinition, and putchar does exist in libc.so