Fix compilation of the vector tests.
__cxxabiv1::__cxa_vec_new and __cxxabiv1::__cxa_vec_delete expect the
constructor and destructor methods to have a void* return type instead
of a void return type.

Change the tests constructor and destructor methods to match, and
return the input void* as their output.

From the ARM ABI section 3.1:
This ABI requires C1 and C2 constructors to return this (instead of
being void functions) so that a C3 constructor can tail call the C1
constructor and the C1 constructor can tail call C2.

Change-Id: Icb389c10feeae32f8d8e1ac45d78c96aef780992