Use dl_unwind_find_exidx on android to find the unwind table.
This was turned off for the GCC build. Instead, make this file check
if we're using libgcc, and use the GCC method if we are, and use
the system one instead on android.

This splits up the #define for dladdr and dl_unwind_find_exidx into
two defines to make things more clear.

By default, say we're using libgcc when not compiling with clang, but
this can be overridden if desired.

Change-Id: I6aae3d9895fd2d0ef470c99fac8119b16609333f