Do __cxa_begin_cleanup before set_registers().
In the case where we found an exception handler in phase 1,
we are doing set_registers() without doing __cxa_begin_cleanup,
then returning _URC_INSTALL_CONTEXT.

In the other case, where we did not find an exception handler,
we do __cxa_begin_cleanup first. Make the common case where
we found a handler match this behaviour so that __cxa_end_cleanup
will work.

Change-Id: If255889c2b87eae10a13a4c75e0ff9abf0a7d27a