Drop the thumb bit from the IP when finding the EHTable.
In UnwindCursor::setInfoBasedOnIPRegister, we subtract one
from the pc to keep the pc from going beyond the end of the
method.

On arm if in thumb mode, subtracting one only removes the
thumb and leaves us in the same position. So we clear the thumb
bit before subtracting to ensure we move the instruction pointer
to the previous instruction.

This fixes the last failing libc++abi test.

Change-Id: I000fb3a698db77caa4950532c20aab4b0fa486f8