android/support: Implement wcstoXX() functions properly.
This patch implements wcstol(), wcstoul(), wcstod(), ... for the
Android support library. This uses a copy of Musl's 1.1 integer
and floating point scanning functions.

IMPORTANT: The x86 implementation of 'scalbnl()' is broken, at least
in the Gingerbread x86 SDK system image, so add a correct
re-implementation to ensure correct results of the floating-point
scanning routines.

Change-Id: I84fbf33be7f33f1a7e7fbefaba3752285ce71f4f