Remove operator new/delete definitions when using libc++abi.
libc++abi already exports these implementations for these operators as
weak symbols. Having libc++ provide them as well is at best confusing
and at worse an ODR violation during a static link. The libc++
definitions used to be compiled out when libc++abi was built but

http://llvm.org/viewvc/llvm-project?view=revision&revision=160790

removed the preprocessor conditional. Due to the nature of weak
symbols as well as dynamic symbol lookup ordering this bug is hard
to test for and has gone unnoticed.

Change-Id: I732ffecba2e4a8690902aec33f168a9001935a5b